requirements/requirements.txt
amqp==2.3.2
aniso8601==3.0.2
asgiref==2.3.2
async-timeout==3.0.0
attrs==18.2.0
autobahn==18.9.2
Automat==0.7.0
billiard==3.5.0.4
celery==4.2.1
channels==2.1.3
constantly==15.1.0
daphne==2.2.2
Django==2.1.1
django-annoying==0.10.4
django-cors-headers==2.4.0
django-environ==0.4.5
django-extensions==2.1.2
django-graphql-jwt==0.1.13
django-gravatar2==1.4.2
django-js-asset==1.1.0
graphene==2.1.3
graphene-django==2.2.0
graphene-file-upload==1.0.0
graphql-core==2.1
graphql-relay==0.4.5
gunicorn==19.9.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
kombu==4.2.1
Pillow==5.2.0
promise==2.1
PyHamcrest==1.9.0
PyJWT==1.6.4
python-slugify==1.2.6
pytz==2018.5
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
Twisted==18.7.0
txaio==18.8.1
typing==3.6.6
Unidecode==1.0.22
vine==1.1.4
zope.interface==4.5.0
